Authorize virker stadig ikke ud fra JWT token

+++ WeatherReadingsAPI/WeatherReadingsAPI/Controllers/AccountController.cs
@@ -57,7 +57,7 @@ namespace WeatherReadingsAPI.Controllers
 
         // GET: api/Account/5
         [HttpGet("{id}", Name = "Get")]
-        public async Task<ActionResult<UserDto>> Get(int id)
+        public async Task<ActionResult<UserDto>> Get(long id)
         {
             var user = await _context.User.FindAsync(id);
             if (user == null)
